created svc file for get/save user
updated user-register.mjs
